refactor RequestIdentifier and RequqestIdentification to separate file
and add doc comments
