+++ sidecar/src/service/mod.rs
@@ -3,10 +3,12 @@
 
 pub use instance_id::InstanceId;
 pub use queue_id::QueueId;
+pub use request_identification::{RequestIdentification, RequestIdentifier};
 pub use runtime_metadata::RuntimeMetadata;
 pub use serialized_tracer_header_tags::SerializedTracerHeaderTags;
 
 mod instance_id;
 pub mod queue_id;
+mod request_identification;
 mod runtime_metadata;
 mod serialized_tracer_header_tags;
